+++ GANs/esrgan/utils/__init__.py
@@ -1,10 +1,10 @@
 from __future__ import absolute_import
 import os
 import sys
-common_utils_path = os.path.join(
-    "/", *os.path.abspath(__file__).split("/")[:-4], "utils")
+common_utils_path = os.path.abspath(
+    os.path.join(os.path.dirname(__file__), '..', '..', '..', 'utils'))
 sys.path.append(common_utils_path)
-from yaml_wrapper import read_yaml
-from misc import AttrDict
-from comm import CommunicatorWrapper
-from gan_losses import RelativisticAverageGanLoss, GanLoss
+from neu.yaml_wrapper import read_yaml
+from neu.misc import AttrDict
+from neu.comm import CommunicatorWrapper
+from neu.gan_losses import RelativisticAverageGanLoss, GanLoss
