+++ capsule_net/_checkpoint_nnp_util.py
@@ -20,5 +20,5 @@ import sys
 # Import save/load_checkpoint from utils
 sys.path.append(os.path.join(
     os.path.dirname(__file__), '../utils/'))
-from checkpoint_util import save_checkpoint, load_checkpoint
-from save_nnp import save_nnp
+from neu.checkpoint_util import save_checkpoint, load_checkpoint
+from neu.save_nnp import save_nnp
