+++ utils/neu/comm.py
@@ -15,27 +15,34 @@
 import nnabla as nn
 
 
+def create_float_context(ctx):
+    from nnabla.ext_utils import get_extension_context
+    ctx_float = get_extension_context(ctx.backend[0].split(':')[
+                                      0], device_id=ctx.device_id)
+    return ctx_float
+
+
 class CommunicatorWrapper(object):
     def __init__(self, ctx):
         try:
             import nnabla.communicators as C
             comm = C.MultiProcessDataParallelCommunicator(ctx)
+            comm.init()
+            self.n_procs = comm.size
+            self.rank = comm.rank
+            self.comm = comm
         except Exception as e:
             print(e)
             print('No communicator found. Running with a single process. If you run this with MPI processes,'
                   ' all processes will perform totally same.')
             self.n_procs = 1
-            self.rank = 0 if ctx.context == "cpu" else int(ctx.device_id)
-            self.ctx = ctx
+            self.rank = 0
             self.comm = None
-            return
 
-        comm.init()
-        self.n_procs = comm.size
-        self.rank = comm.rank
+        if self.n_procs > 1:
+            ctx.device_id = str(self.rank)
         self.ctx = ctx
-        self.ctx.device_id = str(self.rank)
-        self.comm = comm
+        self.ctx_float = create_float_context(ctx)
 
     def all_reduce(self, params, division, inplace):
         if self.n_procs == 1:
